Implemented user-specific rulesets. Loading and unloading of the rules are accomplished through the modfilter bash script. It should be placed in /bin/. The script uses iptables-restore -n to load the rulesets. iptables-restore requires root access, and to run it you should add it to the sudoers file like this;   %authipt ALL = NOPASSWD: /sbin/iptables-restore
Here, "authipt" is the group with the users that are to log in and authenticate. The
user-specific rulesets should be placed in the files uprules and downrules in directory
/etc/authipt/users/USERNAME/. Make sure the "downrules" script undoes all the rules that
"uprules" commit.

Within the uprules and downrules, you can use the strings $userip, $username and $userpid, which
will be expanded to the IP and username of the authenticated user, and the PID of the authipt
process. The $username and $userpid is useful if you want to add comments to your rules.